Add validations

+++ app/Http/Controllers/V1/Cecy/DetailPlanificationController.php
@@ -19,7 +19,9 @@ use App\Http\Requests\V1\Cecy\ResponsibleCourseDetailPlanifications\DestroysDeta
 use App\Http\Requests\V1\Cecy\ResponsibleCourseDetailPlanifications\GetDetailPlanificationsByPlanificationRequest;
 use App\Http\Requests\V1\Cecy\ResponsibleCourseDetailPlanifications\RegisterDetailPlanificationRequest;
 use App\Http\Requests\V1\Cecy\ResponsibleCourseDetailPlanifications\ShowDetailPlanificationRequest;
+use App\Http\Requests\V1\Cecy\ResponsibleCourseDetailPlanifications\UpdateDetailPlanificationRequest as UpdateDetailPlanification;
 use App\Http\Requests\V1\Cecy\DetailPlanifications\UpdateDetailPlanificationRequest;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class DetailPlanificationController extends Controller
 {
@@ -71,35 +73,47 @@ class DetailPlanificationController extends Controller
     /**
      * Store a detail planification record
      */
-    // Algo mal pasa con el request
     public function registerDetailPlanification(RegisterDetailPlanificationRequest $request)
     {
-        // $loggedInInstructor = Instructor::where('user_id', $request->user()->id)->get();
+        $loggedInInstructor = Instructor::where('user_id', $request->user()->id)->first();
+        if (!$loggedInInstructor) {
+            return response()->json([
+                'data' => '',
+                'msg' => [
+                    'summary' => 'Failed',
+                    'detail' => 'No es instructor o no se encuentra registrado',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
 
         $planification = Planification::find($request->input('planification.id'));
-        // $responsibleCourse = $planification->reponsibleCourse();
+        $responsibleCourse = $planification->responsibleCourse()->first();
 
-        // validar que la planification le pertenezca al docente logeado
-        // if ($loggedInInstructor->id !== $responsibleCourse->id) {
-        //     return response()->json([
-        //         'msg' => [
-        //             'summary' => 'No le pertenece esta planificación',
-        //             'detail' => '',
-        //             'code' => '400'
-        //         ]
-        //     ], 400);
-        // }
+        if ($loggedInInstructor->id !== $responsibleCourse->id) {
+            return response()->json([
+                'data' => '',
+                'msg' => [
+                    'summary' => 'Failed',
+                    'detail' => 'No le pertece esta planificación',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
 
         //validar que la planification ha culminado
-        // if ($planification->state()->first()->code === State::CULMINATED) {
-        //     return response()->json([
-        //         'msg' => [
-        //             'summary' => 'La planificación ha culminado.',
-        //             'detail' => '',
-        //             'code' => '400'
-        //         ]
-        //     ], 400);
-        // }
+        if (
+            $planification->state()->first()->code === State::CULMINATED ||
+            $planification->state()->first()->code === State::NOT_APPROVED
+        ) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'La planificación ha culminado o no fue aprovada.',
+                    'detail' => '',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
 
         $state = Catalogue::firstWhere('code', State::TO_BE_APPROVED);
         $classroom = Classroom::find($request->input('classroom.id'));
@@ -118,7 +132,6 @@ class DetailPlanificationController extends Controller
 
         $detailPlanification->ended_time = $request->input('endedTime');
         $detailPlanification->started_time = $request->input('startedTime');
-        $detailPlanification->plan_ended_at = $request->input('planEndedAt');
 
         if ($request->has('observations')) {
             $detailPlanification->observations = $request->input('observations');
@@ -157,23 +170,49 @@ class DetailPlanificationController extends Controller
     /**
      * Update a detail planification record
      */
-    // DetailPlanificationController
-    public function updateDetailPlanification(UpdateDetailPlanificationRequest $request, DetailPlanification $detailPlanification)
+    public function updateDetailPlanification(UpdateDetailPlanification $request, DetailPlanification $detailPlanification)
     {
-        $loggedInstructor = Instructor::where('user_id', $request->user()->id)->get();
-        $planification = Planification::find($request->input('planification.id'));
-        $responsibleCourse = $planification->reponsibleCourse();
-
-        if ($loggedInstructor->id !== $responsibleCourse->id) {
+        // username:1095554529 ->instructor
+        // username:1004242743 ->instructor
+        $loggedInInstructor = Instructor::where('user_id', $request->user()->id)->first();
+        if (!$loggedInInstructor) {
             return response()->json([
+                'data' => '',
                 'msg' => [
-                    'summary' => 'No le pertece esta planificación',
-                    'detail' => '',
+                    'summary' => 'Failed',
+                    'detail' => 'No es instructor o no se encuentra registrado',
                     'code' => '400'
                 ]
             ], 400);
         }
 
+        $planification = Planification::find($request->input('planification.id'));
+        $responsibleCourse = $planification->responsibleCourse()->first();
+
+        if ($loggedInInstructor->id !== $responsibleCourse->id) {
+            return response()->json([
+                'data' => '',
+                'msg' => [
+                    'summary' => 'Failed',
+                    'detail' => 'No le pertece esta planificación',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
+
+        //validar que la planification ha culminado
+        if (
+            $planification->state()->first()->code === State::CULMINATED ||
+            $planification->state()->first()->code === State::NOT_APPROVED
+        ) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'La planificación ha culminado o no fue aprovada.',
+                    'detail' => '',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
 
         $classroom = Classroom::find($request->input('classroom.id'));
         $days = Catalogue::find($request->input('day.id'));
@@ -196,6 +235,7 @@ class DetailPlanificationController extends Controller
 
         $detailPlanification->save();
 
+
         return (new ResponsibleCourseDetailPlanificationResource($detailPlanification))
             ->additional([
                 'msg' => [
@@ -209,9 +249,48 @@ class DetailPlanificationController extends Controller
     /**
      * Delete a detail planification record
      */
-    // DetailPlanificationController
     public function deleteDetailPlanification(DeleteDetailPlanificationRequest $request, DetailPlanification $detailPlanification)
     {
+        $planification = $detailPlanification->planification()->first();
+        $responsibleCourse = $planification->responsibleCourse()->first();
+
+        $loggedInInstructor = Instructor::where('user_id', $request->user()->id)->first();
+        if (!$loggedInInstructor) {
+            return response()->json([
+                'data' => '',
+                'msg' => [
+                    'summary' => 'Failed',
+                    'detail' => 'No es instructor o no se encuentra registrado',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
+
+        if ($loggedInInstructor->id !== $responsibleCourse->id) {
+            return response()->json([
+                'data' => '',
+                'msg' => [
+                    'summary' => 'Failed',
+                    'detail' => 'No le pertece esta planificación',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
+
+        //validar que la planification ha culminado
+        if (
+            $planification->state()->first()->code === State::CULMINATED ||
+            $planification->state()->first()->code === State::NOT_APPROVED
+        ) {
+            return response()->json([
+                'msg' => [
+                    'summary' => 'La planificación ha culminado o no fue aprovada.',
+                    'detail' => '',
+                    'code' => '400'
+                ]
+            ], 400);
+        }
+
         $detailPlanification->delete();
 
         return (new ResponsibleCourseDetailPlanificationResource($detailPlanification))
@@ -261,9 +340,8 @@ class DetailPlanificationController extends Controller
     }
 
     /**
-     * Delete a detail planification record
+     * Delete some detail planification records
      */
-    // DetailPlanificationController
     public function destroysDetailPlanifications(DestroysDetailPlanificationRequest $request)
     {
         return 'works!';
